put all configs into single header + remove config utils from pub api
