+++ src/aktualizr_info/aktualizr_info_config.cc
@@ -1,5 +1,7 @@
-#include <libaktualizr/config_utils.h>
 #include "aktualizr_info_config.h"
+#include "utilities/config_utils.h"
+#include "logging/logging.h"
+
 
 AktualizrInfoConfig::AktualizrInfoConfig(const boost::program_options::variables_map& cmd) {
   // Redundantly check and set the loglevel from the commandline prematurely so
