+++ src/libaktualizr/config/CMakeLists.txt
@@ -1,4 +1,3 @@
-#set(HEADERS config.h)
 set(SOURCES config.cc)
 
 add_library(config OBJECT ${SOURCES})
