+++ src/libaktualizr/package_manager/packagemanagerconfig.cc
@@ -3,7 +3,9 @@
 #include <boost/algorithm/string/split.hpp>
 #include <boost/log/trivial.hpp>
 
-#include <libaktualizr/packagemanagerconfig.h>
+#include <libaktualizr/config.h>
+
+#include "utilities/config_utils.h"
 
 void PackageConfig::updateFromPropertyTree(const boost::property_tree::ptree& pt) {
   for (const auto& cp : pt) {
