+++ src/libaktualizr/telemetry/telemetryconfig.cc
@@ -1,5 +1,6 @@
-#include <libaktualizr/telemetryconfig.h>
-#include <libaktualizr/config_utils.h>
+#include <libaktualizr/config.h>
+
+#include "utilities/config_utils.h"
 
 void TelemetryConfig::updateFromPropertyTree(const boost::property_tree::ptree& pt) {
   CopyFromConfig(report_network, "report_network", pt);
