+++ src/libaktualizr/utilities/config_utils.h
@@ -81,6 +81,19 @@ inline void CopyFromConfig(KeyType& dest, const std::string& option_name, const 
   }
 }
 
+template<>
+inline void CopyFromConfig(StorageType& dest, const std::string& option_name, const boost::property_tree::ptree& pt) {
+  boost::optional<std::string> value = pt.get_optional<std::string>(option_name);
+  if (value.is_initialized()) {
+    std::string storage_type{StripQuotesFromStrings(value.get())};
+    if (storage_type == "sqlite") {
+      dest = StorageType::kSqlite;
+    } else {
+      dest = StorageType::kFileSystem;
+    }
+  }
+}
+
 template <>
 inline void CopyFromConfig(CryptoSource& dest, const std::string& option_name, const boost::property_tree::ptree& pt) {
   boost::optional<std::string> value = pt.get_optional<std::string>(option_name);
@@ -121,50 +134,5 @@ inline void WriteSectionToStream(T& sec, const std::string& section_name, std::o
   sec.writeToStream(os);
   os << "\n";
 }
-
-class BaseConfig {
- public:
-  virtual ~BaseConfig() = default;
-  void updateFromToml(const boost::filesystem::path& filename) {
-    LOG_INFO << "Reading config: " << filename;
-    if (!boost::filesystem::exists(filename)) {
-      throw std::runtime_error("Config file " + filename.string() + " does not exist.");
-    }
-    boost::property_tree::ptree pt;
-    boost::property_tree::ini_parser::read_ini(filename.string(), pt);
-    updateFromPropertyTree(pt);
-  }
-  virtual void updateFromPropertyTree(const boost::property_tree::ptree& pt) = 0;
-
- protected:
-  void updateFromDirs(const std::vector<boost::filesystem::path>& configs) {
-    std::map<std::string, boost::filesystem::path> configs_map;
-    for (const auto& config : configs) {
-      if (!boost::filesystem::exists(config)) {
-        continue;
-      }
-      if (boost::filesystem::is_directory(config)) {
-        for (const auto& config_file : Utils::getDirEntriesByExt(config, ".toml")) {
-          configs_map[config_file.filename().string()] = config_file;
-        }
-      } else {
-        configs_map[config.filename().string()] = config;
-      }
-    }
-    for (const auto& config_file : configs_map) {
-      updateFromToml(config_file.second);
-    }
-  }
-
-  static void checkDirs(const std::vector<boost::filesystem::path>& configs) {
-    for (const auto& config : configs) {
-      if (!boost::filesystem::exists(config)) {
-        throw std::runtime_error("Config directory " + config.string() + " does not exist.");
-      }
-    }
-  }
-
-  std::vector<boost::filesystem::path> config_dirs_ = {"/usr/lib/sota/conf.d", "/etc/sota/conf.d/"};
-};
 
 #endif  // CONFIG_UTILS_H_
